Remove max template depth
Summary: This dates all the way back to the original open-sourcing of
HPHPc. It's been bumped a couple times since then -- and we're running
into it again on Ubuntu 14.10 and Debian 8. It looks like the C++11 spec
allows this to be as high as 1024, and the g++ default is 900, so let's
just do that and be done with it.

Reviewed By: @sgolemon

Differential Revision: D1977061