Modify ColAdd{New,}ElemC spec/behavior to not require isCollection check
Summary: The spec claimed they executed certain expressions if the input base
was a collection, and otherwise fataled.  However, if the input was a
collection, they both subsequently assumed it was only possibly collections of
certain types, in violation of the specified behavior (e.g. you can crash
ColAddElem by passing it a Vector base).  If we're going to assume things like
that, let's just assume it's a collection in the first place, also.

(We should probably also move those decrefs to the TC so they can possibly be
eliminated by refcount opts, but that's something else.)

Reviewed By: @binliu19

Differential Revision: D1981508