Have error messages point to where a constraint is introduced
Summary: When type constraints are involved, our errors would point to
the definitions of the constraints, but not the places where the
constrained generic type actually gets used. This made it difficult both
to locate the error and to suppress them with HH_FIXMES in locations
that made sense. This commit fixes that.

There are two things that make it tricky for us to give precise positions
for error messages involving constrained generics: First, we delay the
constraint checking until we have inferred concrete types for the generics,
and second, generic types (being instantiated to `Tunresolved`s) can grow
as the typing of the function progresses. Ideally, when encountering a
constraint violation, we would point to where the generic type got
instantiated (which is also the position where the constraint gets registered
as a `todo`), as well as the point where the problematic type got added to
the `Tunresolved` -- at least one of these two positions must be changed.
Unfortunately, our existing implementation did not track either of these
positions.

My first approach involved having Tunresolved types tack on a `Reason.t`
that recorded the position where a type got added to a Tunresolved list.
However, we create Tunresolveds in many places, and having each one
tack on a `Reason.t` makes for rather spewy errors. So my current approach
just tracks the places where instantiation occurs. This is somewhat less
precise, but it is a vast improvement over the current situation, and it seems
to work well enough -- I am now able to locate the source of errors that
were previously mysterious.

It's still a bit of an awkward situation because our existing error
system limits us to wrapping our unification error messages with more
detail. Suppose we are trying to unify `T super A` with `B`, and B is
not a supertype of A. `Unify.unify` will create an error of the form

   [some position] This is an object of type A
   [some position] It is incompatible with an object of type B

Firstly, it talks about values when we really want to talk about types,
and it emits strings like "an object of type ...", which imply
subtyping; this is unfortunate because we really want to talk about
supertype constraints here.

Ideally, our error message would be something like

  A constraint introduced here is violated
  This is the constrained type
  It is constrained to be a supertype of `A`
  An object of type `B` is not a supertype of `A`

But that would require changing the unification error message depending
on the caller of `Unify.unify`, and that's more work than I am willing
to do now.

Reviewed By: @elgenie

Differential Revision: D1982844