Don't ignore args RegSets in vasm-llvm
Summary: rVmSp is going to disappear from kCrossTraceRegs soon. This diff
modifies the llvm backend to inspect the args RegSets on relevant vasm
instructions, only passing arguments for the PhysRegs present in the
set. Codegen should be unaffected for now, and when rVmSp disappears from
kCrossTraceRegs the llvm backend should automatically stop passing it to
stub/helpers/etc.

Reviewed By: @jdelong

Differential Revision: D1978003