Ability to attach metadata to a stack frame
Summary: Introduce a set_frame_metadata() function that attaches metadata to the calling function. The metadata are accessible via debug_backtrace() with a special flag.

This feature can be used to supply additional info to debugging and profiling tools, such as Xenon.

Future optimization work outside of scope of this diff:

- declare 86metadata named local in methods which call HH\set_frame_metadata() to avoid construction of dynamic variable environment
- optimize JIT CallBuiltin to HH\set_frame_metadata() with a simple local assignment to avoid extension call

Reviewed By: @billf

Differential Revision: D1778113