Skip instantiation if there is nothing to substitute
Summary: While looking into the perf implications of D1982724 I noticed how expensive it
is to recursively deconstruct a type. With this intuition I noticed an obvious
case where we can improve perf by avoiding deconstructing the type.

If there is nothing to substitue for instantiation we can exit early and avoid
the cost of deconstructing the type. This seems to reduce the time we spend in
the Type-check phase by ~10%

Reviewed By: @int3, @jwatzman

Differential Revision: D1987578