Consider T as X to be subtype of X in more cases
Summary: Currently we only check the as constraint of a generic in subtyping of the super type we are comparing to is a primitive or a Tapply. We basically always want to utilize the as constraint of a generic if we can when doing subtyping. This manifested as a bug when the subtype we compared against was a Tabstract.

I fixed this bug by moving this logic to the bottom of the match in subtype_with_uenv. This has to go below all the stuff we do with the uenv for Toption. I don't fully understand why, but if it appears above it then various test cases start failing.

Reviewed By: @jwatzman, @elgenie

Differential Revision: D1976546