Cache time zone validity and add lock for calling timelib on Mac.
date_default_timezone_set() is called once for every WordPress request in
stock WordPress 4.1.1. Cache the time zone validity for better performance.

A lock is needed when calling timelib_timezone_id_is_valid() because
setlocale() on Mac is not thread safe. When the first argument to setlocale
is LC_CTYPE, loadlocale() will call __setrunelocale() to modify the global
locale object and double free can happen and crash.

http://www.opensource.apple.com/source/Libc/Libc-997.90.3/locale/FreeBSD/setrunelocale.c

Part of #4444.