redis responseDouble returns false on error, not null
I am thinking about why there is null in the first place. I thought that this is for a reason.
My guess is that the one who implemented this looked at php's implementation
and saw the `return;` call. But that isn't called actually.

So while not being 100% sure that this is correct for all cases, and having
not the best test coverage on redis (because we don't import the php-src
tests) I think this is most likely correct.

Fixes #5112

Test Plan:
new test && slow/ext_redis/
Note that these tests probably get skipped on sandcastle