Use movups instead of movdqu in codegen
Summary: It's nearly identical and it's a byte smaller. The only difference,
according to some random dude on the internet[1], is that there might be an
extra cycle or two in arithmetic operations due to forwarding data between
different ports in the CPU. That could only matter if we actually did math to
packed quadwords in xmm registers after loading, so this should be strictly
better.

[1] http://www.gamedev.net/blog/615/entry-2250281-demystifying-sse-move-instructions/

Reviewed By: @edwinsmith

Differential Revision: D1991228