simpleProp depends on clsSpec but doesn't constrain on it
Summary: simpleProp minstrs check that the class of the base doesn't
have unknown property offsets. Thus, it depends on knowing the class of
the base object, but it doesn't report that via constrainBase.
getPropertyOffset also uses rtt information from NormalizedInstruction's
inputs, which has no bearing on what the proven or predicted type of the
base object currently is.

Reviewed By: @swtaarrs

Differential Revision: D1973398