Simplify the DefLabel case in GVN
Summary: Previously we handled DefLabel separately from the other
opcodes for which we just checked supportsGVN. This led to at least one
bug where we were GVN-ing an instruction that was visible through a
DefLabel but that didn't supportsGVN. This fix incorporates a lot of the
previous logic for being able to GVN DefLabels with some
tweaks/additions. It changes supportsGVN to recursively check the inputs to a
DefLabel to verify that they also supportsGVN. The recursive nature of the check
is important so that the analysis can safely see through multiple layers of
DefLabels. Treating DefLabels "the same" as other GVN-able instructions
also required the NameTable's Hasher and Comparator to account for DefLabel's
uniqueness.

Reviewed By: @swtaarrs

Differential Revision: D1984303