Touch file when something changes
Summary: This will be used to inform HHVM that "something changed", in
order to facilitate caching in the `\HH\Client\typecheck()` function
introduced in D1940190. This mechanism is deliberately stupidly simple:
just `touch` a file in a known location, with every `hh_server` using
the same file. The idea here isn't to be super precise, just to deal
with the common hot case of "nothing changed" efficiently.

Reviewed By: @elgenie

Differential Revision: D1960376