redis responseDouble returns false on error, not null
Summary: I am thinking about why there is null in the first place. I thought that this is for a reason.
My guess is that the one who implemented this looked at php's implementation
and saw the `return;` call. But that isn't called actually.

So while not being 100% sure that this is correct for all cases, and having
not the best test coverage on redis (because we don't import the php-src
tests) I think this is most likely correct.

Fixes #5112

Closes https://github.com/facebook/hhvm/pull/5174

Reviewed By: @oyamauchi

Differential Revision: D1991553