Turn ActRec flags into an enum to make it a little easier to extend
Summary: Since InResumed and FromFPushCtor are mutually exclusive, this
effectively gives us one more state for free.  This diff doesn't use the state,
it just refactors things so it's easy to use that in something else I'm trying
(but it seems worth landing even if that doesn't work because we might want to
use it for something else).  Also removes some unused functions and constants
and stuff.

Reviewed By: @swtaarrs

Differential Revision: D2002544