Don't require storing null to ActRec::m_invName at non-magic callsites
Summary: Magic callsites need an invName, and need to know if it was a
"non-magic" call of the __call method by whether the invName is null.  No other
callees need the invName field until after the ActRec becomes live, and then
they only need it if AttrMayUseVV is set.

This changes magic dispatch to set some flags on the ActRec, and changes magic
callees to use those flags to determine whether the call was "magic".  This
means non-magic callsites never need to store null into the invName field.
Then in function prologues for non-magic functions with AttrMayUseVV, we now
have to null out the invName field if we didn't store an ExtraArgs field, so
that later portions of the function that might create a VarEnv don't get
confused.

This also means that it's illegal to use things like ActRec::getVarEnv without
checking if the func has AttrMayUseVV first, because otherwise you could read
garbage.

Reviewed By: @jano

Differential Revision: D2013101