Don't use unpack operator for MethCallerHelper if there are no args
Summary: In {D1961249} we used the unpack operator to have meth_caller() work with
functions with any number of args. Unfortunately the unpack operator does not
work for functions that have references as arguments. This will break some
current callers of meth_caller(), so we special case the no-arg version to
invoke the method without using unpack.

Reviewed By: @elgenie

Differential Revision: D1998106