Remove 3 second delay on small project startup time
Summary: When the server wasn't running, we'd start it up, and then
sleepl for 3 seconds before trying again, to make sure it had begun
initalizing. This is silly: the total startup time for small projects is
much less than 3 seconds, and we shouldn't be polling in the first
place. Instead, we can have the server just send us a signal to tell us
when it's in the appropriate state.

This entire code is polling based, and generally poorly factored. This
fixes both the worst offender and the most easily fixed (since we can
easily send a command-line argument at startup). Ideally we wouldn't
poll during initalization either, but that's harder since we can't
actually tell the server what to do once it gets starting.

FWIW I'm not super happy with this, I'd like something more
general, but this solves a pretty bad problem, and the general
solution would involve a rearchitecting of the way the client
and server communicate, more or less.

Reviewed By: @int3, @elgenie

Differential Revision: D2003464