Use SmartPtr in libxml
Summary: Replace uses of raw ResourceData and ObjectData with SmartPtr<T> and Object respectively.
Use makeSmartPtr instead of newres.
Remove as many uses of get() as possible.

Use RequestEventHandler to clear s_NodeMap instead of InitFiniNode so that the raw ObjectData pointers can be replaced with Objects and that proper ref counting will be performed if an Object is removed from the map.  On shutdown, the entire map can be dropped since it no longer matters what the ref counts are.

I've removed the explicit ref counting for Files that were stored in XML contexts and deleted by the libxml_streams_IO_close callback.  In ext_domdocument and ext_simplexml, the XML context is created and destroyed locally, so it does not need to take ownership of the File.  For these extensions, the context just contains a pointer to a locally owned File.  In ext_xmlreader, the XML context is owned by the native data in a PHP object.  I've moved the file ownership into the native data also and just stored a pointer to it in the XML context.

ext_libxml is more involved since there is no object to store the File in.  I've added a map of (id -> File) to the LibXmlRequestData object.  The XML contexts created in this file reference an id from this map.  The map is cleared on request init and shutdown.  Individual Files will also be closed and deleted by the libxml_streams_IO_close callback.

Reviewed By: @edwinsmith

Differential Revision: D1799352