Remove one of DefInlineFP's StkPtrs
Summary: We never inline calls that have other calls in their FPI region, so
even before we eliminate the ReDefSP/ResetSP instructions associated with
inlined calls it's no longer possible for the return StkPtr to be different
from the StkPtr after all the args are stored.  Also fix some comments that
still referred to SpillStack instructions that don't exist anymore.

Reviewed By: @ottoni

Differential Revision: D2013948