Fix handling of `super` constraints
Summary: My initial implementation raised a error for the following, complaining
that B is not a supertype of C:

  class A {}
  class B extends A {}
  class C extends A {}
  class Foo<T> {
    public function bar<Tu super T>(Tu $x): Tu {}
  }
  function f(Foo<C> $x, B $y): A {
    return $x->bar($y);
  }

It shouldn't be an error, however, because the constraint is satisfied
if `Tu = A`. This fixes it (by reverting to @paroski's original
implementation), and adds an explanatory comment.

The `super_type` and `sub_type` distinction is now a bit of a misnomer,
since we don't want to toggle `grow_super` when checking super
constraints (it's only toggled when checking contravariance now). I've
replaced it by `invert_grow_super` instead.

Reviewed By: @jwatzman, @elgenie

Differential Revision: D2007453