Introduce Type "Phases" By Utilizing GADTs
Summary: Currently we represent types in both the Decl Phase and Local Typing Phase with the same variant type Typing_defs.ty. However there are types that should only exist in the Decl Phase but not the Local Typing Phase and vice versa. This leads to assert false branches in pattern matching. More importantly there is some ceremony that should be performed when moving a type from the Decl Phase to the Local Typing Phase, but this is not centralized anywhere and instead relies on copy pasta throughout various parts of the typing code to do the right thing.

To tackle both issues I worked on introducing a phantom type to Typing_defs.ty to denote what phase it is in. In addition to the phantom type I made the ADT for Typing_defs.ty into a GADT that informs the compiler that certain types only exist in certain phases. For instance ##Tobject## is not a valid type declaration in Hack and thus should not appear in code related to the Decl Phase, however it can show up during the Local Typing Phase. The compiler can then use this information to identify unneeded branches in pattern matching as well as making it obvious you forgot to go through the proper ceremony for using a Decl Type for a function that expects a Local Type.

This diff introducing a function Typing_utils.localize that translates between the phases. At the moment it utilizes Obj.magic to do the translation. The rest of the diff consistence of determining what functions apply to which phase and adding types and calls to localize until the compiler is happy. I've surely added more calls to localize than is necessary, but I wanted this diff to be a starting point.

Reviewed By: @jwatzman

Differential Revision: D1982724