OCaml must be at least 4.01
Summary: As of D1982724, we require OCaml version 4.00. Several distros,
notably Debian stable, only have 3.12, which is what we've historically
supported. But the changes in D1982724 are too important and too good a
cleanup to revert and not do -- so we'll have to live with this.

This diff bumps up the minimum version of OCaml to 4.01 (we technically
work with 4.00 I think, but no distro has that, they are all at 4.01 at
least). For the older distros that have 3.12, if you want to build HHVM,
you'll have to install your own OCaml package. IMO, that's fine --
people on those distros already need a newer version of more basic
things like g++ anyways. I'll write up directions on the wiki on how to
do it step by step.

This diff also makes OCaml a //requirement// to build HHVM. Hack is an
integral enough part of the ecosystem now IMO that we don't want people
playing around with HHVM -- potentially writing and running Hack
code! -- without also having the typechecker.

Reviewed By: @fredemmott

Differential Revision: D2017280