Redo tagged support to better support ARM64.
This redos tagged pointer support but disables for now the generic
implementation.
For ARMv8, 8bit tags can be supported without anding them out.
For right now armv8 supports up to 48bit virtual address space
but in the future it will support larger space.

This builds for arm64 without any new bugs, it should work for x86_64 but I have not tried it.
For PowerPC, you need to do specializations just as what was done for x86_64 and arm64 or just use the generic versions.