Don't redefine stack pointers for inlined functions
Summary: This turned out to be less difficult than I thought it was: it turns
out there's no problem with letting just letting the spOffset in FrameState
stay relative to the spValue we had outside the callee.  All stack addressing
was ok with the possibility that it's negative already, since it already could
be during a return sequence.  (Also it's not always negative in this situation
anyway: if you enter a region, pop some stuff, then start inlining, we could
still have a positive offset here.)

This diff removes ReDefSP and ResetSP, which were only needed for
inlining-related stuff.  We're down to just RetAdjustStk and AdjustSP that
define stack pointers, which are used in some region-exiting locations that are
still semi-legit (we might want to package them into the instructions that
actually leave the region---at least some of them are unnecessary unless you're
in a resumed function now---then no-IR instructions except DefSP will define a
StkPtr).  But after this diff we should be able to remove the ResetSP in
prepareForHHBCMergePoint.

Reviewed By: @ottoni

Differential Revision: D2021615