Make some hidden FramePtr dependencies explicit in the IR
Summary: Several instructions that leave the trace don't take the FramePtr.
(Code sequences around these used to use SyncABIRegs for this purpose, but that
was removed.)  They generally force a DefInlineFP to remain because they depend
on the StkPtr that comes out of the ReDefSP after it, but D2021615 is taking
that away, so without this diff we'll start eliminating DefInlineFPs that we
shouldn't.

Reviewed By: @swtaarrs

Differential Revision: D2025481