Move Expansion of Typedefs into Localize
Summary: This is a demonstration of what we can do with {D1982724}. I moved our Tapply type to the DeclPhase and introduced a Tclass into the LoclPhase. This allows us to enforce an invariant that you cannot going from a DeclType to a LoclType without expanding type aliases.

When refactoring this code I also introduce another environment that tracks information while expanding a type. I utilize this enviornment to keep track of the typedefs that were expanded while localizing, but down the road we can use this to also detect recursive Taccess, Tvar, or nested Toption.

Depends On D1982724

Reviewed By: @jwatzman

Differential Revision: D1982767