Move Statement::KindOf and Expression::KindOf into Construct
Summary: Both Statement and Expression have a KindOf field. A lot of
dynamic casting goes on in the parser to determine if a Construct is
either a Statement or an Expression. By lifting these two KindOf fields
into Construct, a lot of those checks either become cheaper or don't
need to be done at all.

Reviewed By: @jdelong

Differential Revision: D2016093