Refactor Config::[Get|Bind] functions for ini migration
Summary: In order to facilitate the transition from hdf to ini, this simplifies the
`Config::Get` and `Config::Bind` functions to handle all scenarios. When we
finally migrate, we should theoretically be able to remove the `Hdf` parameter
and move forward with ini only.

All functions now take a `name` parameter, in the Hdf form, which gets converted
to its appropriate ini setting. There is also a parameter as to whether we
prepend `hhvm.` in front of a settings. This is mostly helpful when we do
iterations through settings (e.g., `Config::Iterate` and the current
`for hdf.firstChild()` scenarios), but is also useful for settings like `extension_dir`,
which is a root setting, but with no `hhvm.` in front of it.

Speaking of `for hdf.firstChild()`, these will go away in future diffs, being replaced
with `Config::Iterate()`. In fact, three current diffs D1906717, D1883841 and D1920884
do this change already and they will be rebased upon this refactor.

Reviewed By: @markw65

Differential Revision: D1939332