Replace incorrect uses of AliasAnalysis::must_alias and rename that function
Summary: The AliasAnalysis::must_alias function was a little too confusing, and
accordingly was already being misused in one place.  The main use of it with
kills sets in various places is legitimate, but maybe a better metaphor is
"expanding" an AliasClass into the ALocBits that we know are contained in it.
It only makes sense to do this with AliasClasses that we want to interpret
"exhaustively" (i.e. they refer to /all/ locations contained in the class, not
just some locations in it).

So, uses of must_alias with PureStore instructions were incorrect (only one
existed in refcount-opts to do a drop_support): this could go wrong if we had a
PureStore to a PtrToGen---we would drop support on AUnknown, and the
specification of must_alias was such that this could legally return an ALocBits
with everything set.  Mwilliams was fleshing out must_alias to support more
classes and ran into this bug.  I think this wasn't causing problems yet only
because we don't create generic-enough pointer types to the AliasClasses that
must_alias supported so far.

OTOH, it is ok for us to require PureSpillFrame gives back an exact stack class
for its frame cells, so we can use expand() there.

Reviewed By: @mxw

Differential Revision: D2032007