Fix findDefiningBlock
Summary: findDefiningBlock was too conservative in that it was returning
nullptr in cases where there's clearly a first block where the value
was available.  Most specifically, this was happening if the value was
defined by a branching instruction that was the pre-header of a loop.
This diff fixes this issue by adding some extra dominance checks,
which required passing IdomVectors to findDefiningBlock.

Reviewed By: @jdelong

Differential Revision: D2024351