Ternary expressions should not leak lenv changes
Summary: We should not allow the ternary conditional to affect
subsequent statements.

Remark: Ternary expressions are still somewhat buggy -- they don't
handle assignments in their branches correctly, nor do they check if a
branch is terminal (e.g. which could happen if it invokes e.g.
`invariant_violation()`). To fix this we should either duplicate the
logic in `If` fully, or ban assignments and terminal expressions in
ternary expressions... I'm in favor of the latter. Anyway, these should
be fairly rare, so I'm not too worried.

Reviewed By: @jwatzman, @elgenie

Differential Revision: D2031168