More accurate ReturnEffects
Summary: At a ReturnHook we've already destroyed the frame,
and the hook isn't allowed to look at the stack, so it can
mark both as killed.

At a RetCtrl thats not part of a resumable, there's nothing on
the stack, and the frame has been destroyed, so we can do the
same. The one problem was that the final return from a generator
was marked as not being part of a resumable; I think thats wrong
because in that case we have a null on the stack, and the frame
is still live.

Reviewed By: @jdelong

Differential Revision: D2042740