Handle one-argument lambdas correctly
Summary: Previously, when trying to format a one-argument lambda like
`($x) ==> ...`, we would look at `$x`, conclude that the parentheses
contained an ordinary expression, and then attempt to parse the
remainder of the tokens as expr_remain. This patch fixes it so that we
explicitly look for the `==>` token after the closing paren. If the
lambda arrow is not found, then we fall back to trying to parse `$x` as
an ordinary expression.

This is similar to what the parser does with `try_short_lambda`.

Reviewed By: @jwatzman

Differential Revision: D2045723