Fix some issues in insertAssertions
Summary: - It could insert DbgAssertRefCount instructions that use SSATmps that aren't
    defined.

  - StStk no longer implies the value being stored is not already deallocated,
    so we can't assert values right after StStks.

  - Tweak a comment.

This means we won't have as much assertions on reference counts.  But the new
refcount pass inserts assertions whenever it deletes IncRef or DecRef
instructions, which should be helpful.  I was considering whether we could try
to generate some of these on the values we pass to a StStk during gen time, but
I wonder if it might affect the other optimizations we can do too much and make
some bugs just magically go away when you turn on HHIRGenerateAsserts (because
e.g. something wasn't a dead load anymore), so this seems least bad to me...

Reviewed By: @ottoni

Differential Revision: D2048586