Make irgen Concat support more types
Summary: One of the hot translations in wholecfg is doing an InterpOne to
concatenate two ints (which also causes refcount-opts to pessimize_all, ruining
inc/dec removal elsewhere in the loop right now).  This also gets rid of
ConcatCellCell, makes SetOpL support more cases, and fixes an interpreter/jit
consistency bug in SetOpL about when it was raising warnings.

Reviewed By: @binliu19

Differential Revision: D2050563