Add missing include for runtime.h in emitter.cpp
Summary: Needed for definitions of things like `k_E_DEPRECATED` and
`MixedArray`. This was typically getting pulled in by chance via
`hphp/runtime/ext_zend_compat/hhvm/zend-wrap-func.h` -- but only if ezc
was compile-time enabled, which it isn't on my Mac. (Though this one
doesn't look to be a Mac-only issue.)

Reviewed By: @paulbiss

Differential Revision: D2048088