Remove remaining StkPtr-redefining instruction (RetAdjustStk)
Summary: I was going to just wait to get rid of rbx-syncing-before-vretm to
eliminate this, but it turns out pretty easy (and the rbx-syncing-before-return
is tricky enough that I might do it later or in lockdown), so let's just do
this now.  There is various code in frame-state and alias-class that can be
pulled out after this in a later diff(s).  After this diff, the only
instruction that will have dests that are stack pointers is DefSP (but
technically it's legal to do it with Mov and DefLabel too right now ... unsure
yet if we want to ban that, but probably ...).

Reviewed By: @markw65

Differential Revision: D2051273