Use GeneralEffects for iterator instructions
Summary: The IterEffects thing was too inconvenient, so nothing bothered to
support it.  GeneralEffects can already represent the same information as
IterEffects now, but IterEffects2 will just have too large of a union for the
affected locals in its load/store sets.  (This is better than how it's handled
in all the current client code except store-elim, but we can fix it there by
making the AliasClass unions better later, and not pessimizing in refcount-opts
when we see iterators is probably more important.)

The reason these existed in the first place was because I was originally trying
to plug memory-effects into frame-state/local-effects, which had specific
iterator handling that I was trying to replicate, but that didn't end up part
of things because it broke all sorts of things in other passes that used
frame-state.

Reviewed By: @markw65

Differential Revision: D2053102