Add error for assigning to $this
Summary: We do not have a specific error for attempting to assing to $this. This is a
parse time error in HHVM and we would fail to detect it if the value we were
assigning was of type Tany. Instead of checking the types, we should check the
expression itself to see if its valid.

We could emit an error as early as the naming phase if we wanted, but we emit
errors of the placeholder $_ in the type-check phase so I did the same here to
be consistent.

Reviewed By: @dabek

Differential Revision: D2057270