Async command support in xdebug
Summary: Adds an asynchronous break command, that pauses the debugged request as it is
running.

This is done by adding a polling thread to every request being debugged by a
remote xdebug client.  When the request is busy running, the polling thread
polls the socket connection to the client for commands.  If it gets a break
command, it pauses the request thread and relinquishes control of the socket.

Reviewed By: @paulbiss

Differential Revision: D2043247