Fix some gcc-4.9 compatibility issues
Summary: Update some library requirements, and fix a few
stricter checks in the code.

With this, hhvm compiles, but fails to link due to some iconv
issues, which I didn't try to resolve.

Also,

  fbconfig --platform-all=gcc-4.9-glibc-2.20-fb hphp/hhvm

works, but -r hphp fails, because some ocaml libs that hack
depends on are not up to date, and trying to fix them turned
into the usual rats nest. I'll take a look at those separately.

Reviewed By: @jdelong

Differential Revision: D2059660